notebooks: revert all changes from pavics.ouranos.ca to boreas.ouranos.ca
To avoid confusing users reading the notebooks with 2 hostnames that are
actually aliases of each other (users do not know this).

Recall the original intention to do this is to avoid the need to deploy
the `.ncml` file and all its matching `.nc` files that are pretty large
to all the test servers.  The test servers will simply get them from our
production pavics.ouranos.ca server.  We will need a different way to
achieve this same goal.