Correct datetime for some models
The datetime format of some models data weren't in the good format for the correction manipulation and training. I've managed to correct the issue by transforming the non-standard calendar in numpy.datetime64 which is the desired format. It is now possible to use these models also: CMCC-ESM2, CanESM5, FGOALS-g3, INM-CM4-8 and INM-CM5-0.